// Virtual-Hr Front End/src/app/Models/JobModel.ts
export class JobModel {

    CompanyName: any ;

    JobTitle: any;

    ExpiringDate: any;

    Location: any;

     Discription: any;

     Requirement: any;

    JobType: any;

     Qualification: any;

     YearsOfExperience: any;

    JobField: any;

    Responsibility: any;
    _id: any



}